Added multi threading

+++ raytracing/raytracing/RayMath/hitablelist.h
@@ -15,7 +15,7 @@ public:
 bool HitableList::Hit(const Ray& r, float t_min, float t_max, HitRecord& record) const {
 	HitRecord tempRecord;
 	bool hitAnything = false;
-	double closestSoFar = t_max;
+	float closestSoFar = t_max;
 	for (int i = 0; i < listSize; i++) {
 		if (list[i]->Hit(r, t_min, closestSoFar, tempRecord)) {
 			hitAnything = true;
